Move celery worker under src/main
