+++ src/main/queue.ts
@@ -3,10 +3,10 @@ import { PluginEvent } from '@posthog/plugin-scaffold'
 import * as Sentry from '@sentry/node'
 
 import Client from '../celery/client'
-import Worker from '../celery/worker'
 import { status } from '../status'
 import { IngestEventResponse, PluginsServer, Queue } from '../types'
 import { UUIDT } from '../utils'
+import Worker from './celery/worker'
 import { KafkaQueue } from './ingestion/kafka-queue'
 
 export type WorkerMethods = {
